fix(geometry): split IfcMappedItem processing out of router/processing.rs
72a331a4e's dropped-item reporting fix in the mapped-item loop pushed
processing.rs 23 lines past its module-size ratchet budget. Extract the
IfcMappedItem source resolution (caching, cyclic/depth-bounded recursion,
and merging a RepresentationMap's items into one source-coords mesh) into
a new sibling module, router/mapped_item.rs, following the same pattern
already used for caching.rs, diagnostics.rs, layers.rs, etc. No public
symbol renamed or dropped; IDENTITY_ROW_MAJOR is now pub(crate) so both
modules can share it.

No budget or allowlist value changed.
